pyproject.toml
[project]
name = "spotify-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "spotipy>=2.25.1",
]
